Removed Quotation Marks
Not sure how big of a deal this will be, but from my understanding when you create a Table and Column names using the quotation marks it makes these names case sensitive rather than using the SQL default of being case-insensitive. This will change how queries are ran on the Table. For instance on this table,  the queries for showing all information in the transactionentry table:
SELECT * FROM transactionentry;
SELECT * FROM TRANSACTIONENTRY;
SELECT * FROM TransactionEntry;
SELECT * FROM TrasActionEntry;
and so on, will not work.

The only query that will work to show all the information in the transactionentry table will be:

SELECT * FROM "transactionentry";

This goes for any other queries as well that will be ran on this table, every column name will need to have quotation marks around it and have to match the name exactly case-wise. I think this might have a problem with the foreign key referencing the transaction table because it is R_ID and not r_id and with the quotes I don't know if they will have that relationship. I am just going to fix it by taking all the quotes out, but I wanted to explain why I did it.